Add additional checks

+++ sources/find_product/find_product.cpp
@@ -2,9 +2,8 @@
 
 using std::vector;
 
-unsigned long long
-find_product::computeProductModulo(vector<unsigned long long> array) {
-  unsigned long long product = 1;
+uint64_t find_product::computeProductModulo(vector<uint64_t> array) {
+  uint64_t product = 1;
   for (auto &&a : array) {
     product = product * a % 1000000007;
   }
