+++ sources/find_product/find_product.h
@@ -1,10 +1,11 @@
 #ifndef FIND_PRODUCT_H
 #define FIND_PRODUCT_H
 
+#include <cstdint>
 #include <vector>
 
 namespace find_product {
-unsigned long long computeProductModulo(std::vector<unsigned long long> array);
+uint64_t computeProductModulo(std::vector<uint64_t> array);
 }
 
 #endif
